feat: add search to tables, and move components to their own folders
